v0.1.0 - Logo, Background, Tiles, Blobs
Designed and implemented a logo and a favicon.

Added a background gradient and defined other colors.

Added tiles - elements with glass-like appearance and a noise texture as a part of the background.
They are used as a basic section background for things like nav, project tiles, etc.

Added background blobs - elements that move in the background, making the design more dynamic.
The size, direction, and speed are generated randomly.

Added basic styles.