v0.2.0 - Navigation, Icons, Headings
Added the Navigation section.
It's set to position fixed and changes its height as the user scrolls the page.
The page will always load at the top of the website.

Added styling to the headings.
------------------------------
Mobile Version:

The nav transforms into a square with the logo in the top-right corner of the window.

After the user clicks the nav tile, it will smoothly expand into the full-size vertical navigation containing the section names and the corresponding icons.
Each icon has a unique animation - it's infinitely looped, so if the user keeps the cursor on hover for a long time, the animation will restart after a short time.

The user can collapse the nav by clicking the close button, link or outside the nav.
The logo is smoothly transformed into the close icon on hover.

When the user hovers over the nav, there is a border indicator that changes its state based on the cursor position.
A part of the border will be highlighted (gradient) - the brightness and size will depend on how far the cursor is from the close button.
If the cursor is outside the nav, the border will not be highlighted.