v0.4.0 - About Section, Hero Auto Scroll
Added the About section.
There are a few tiles, each one has a unique, animated icon.
They contain info about the experience, technologies, etc.

Implemented an auto scroll feature for the hero section.
When the user scrolls the page, it will automatically scroll through the hero section - smoothly and slowly.
Similarly, when the user scrolls to the top of the page and reaches the hero section, the page will be automatically scrolled all the way to the top.
When the auto scroll is active, user scroll inputs are disabled.