v0.5.0 - Parallax, Animations, SEO
Changed the way the background blobs are animated.
Instead of utilizing requestAnimationFrame (rAF), the framerate is now determined by the device's screen size.
Blobs are now animated using the framerate that is set based on the device's screen size, instead of rAF, which caused choppy animations on less powerful devices.
The background animation refresh rate on desktops is higher and the mobile animation is set to lower FPS.

Changed the parallax effect to be based on a transition (it's now using a class that is relative to the window scroll state in the hero auto scroll).
The performance is better, since it's just toggling between two states instead of updating the position on a window scroll event.

Fixed the Hero Auto Scroll animation.
Now, the time duration is correct and not affected by the framerate.

Improved the hero auto-scroll behavior.
Perviously, when the user tried to scroll through the hero section, the animation was jittery - especially when swipe-scrolling on mobile.
Now, it's reduced to a minimum. Some browser behaviors are not possible ot overwrite.

Improved the SEO, Accessibility, etc.

Implemented minor design fixes and adjustments.