v0.5.1 - Performance Optimization
Modified the background blobs to be animated at a much lower framerate (the same value for all display sizes) and added a transition, so the animation is smooth, while requiring much less computing.
Simplified and optimized the code, reducing the computing required for the animation.

Removed the backdrop-filter from the tiles completely (on mobile devices).

Adjusted the top padding of the main section.