v0.6.1 - Nav & Auto Scroll Fixes
Fixed the navbar height.

Implemented a nav background gradient variation using CSS relative color syntax.
The nav background was too transparent after the backdrop-filter was removed on mobile, resulting in a bad UX (difficult to read the text).

Fixed the Auto Scroll behavior in the Hero section.
If the user scrolled down and then used the navigation bar to move to the top section (about), the Auto Scroll feature would be initialized due to sections overlapping.