v0.7.0 - Project Tiles & Modal
Changed the way the project modal is displayed.
Now, instead of sliding in from the side of the screen, the modal is initially placed within the tile and expands to cover the entire screen when clicking on a tile.

Changed the tile hover animation to be simpler and shorter.
The active (clicked) tile still has a looped "breating" logo and project title animation.

Fixed the icon transition in the nav.