v0.8.0 - Modal, Project Mockup
The Tile Display (logo and project title) is now a part of the modal.
It will seamlessly transition between the tile / modal mode with the Tile Display always visible.

Fixed a bug where the modal would be misplaced on window resize, especially when resizing quickly.

Added the Project Mockup.
There are 2 modes - desktop and mobile.
The desktop mode resembles a PC monitor and the mobile mode looks like a smartphone.
Each mode has a scrollable project screenshot designed to resemble the respective device.
The default mode is respective to the user's device - mobile for phones/tablets and desktop for PCs/laptops (based on the device's width).
The modes can be toggled.

Fixed and improved many minor design elements, such as opacity and z-index when switching between the tile / modal modes.