v0.8.1 - Mockup Fixes
Changed the mockup toggle design to text.
It changes the text color based on the mode and alters the slash color to indicate transitions between states.

Fixed a behavior where, if the user would switch the mockup rapidly, the image swap would be visible.
Fixed the mockup design on very small deivces (the smartphone buttons were misplaced).