v0.10.0 - Graphic Design Section, Minor Fixes
Updated the TilesSection (ProjectsSection) component to make it possible to use it for different sections with different data.
Changed some parts of the code for correct functionality and better naming.
Changed the even/odd tile number code to adjust the tiles automatically.
Improved the smoothness of the opening tile/modal animation.

Fixed a bug, where if the user clicked a different tile right after closing the previous one, the background color would disappear.

Added the Graphic Design Section.
It contains logo, flyer, icon, and color palette projects.
Each project can be opened in a modal.

Fixed the navigation icons.
The unique animation was not infinite on hover for some icons.
Fixed the color palette icon design, so there is no path gap under the brush on hover.

Changed the "X" modal button slightly to be better visible when on top of the content.

Applied other minor design changes.