v0.10.1 - Modal Backgrond Auto Toggle
Added an auto-toggle feature for the background color in the Graphic Design Project modal.
The code automatically calculates the average lightness of the image (like in HSL) and changes the modal background color if the image is dark/light.
The user can also adjust the background color manually (toggle button).

Applied other minor design changes.