v0.11.0 - Design Section Improvements, Projects
Updated the Graphic Design Section code.
Simplified and removed unnecessary or repeated code, making the project easier to maintain.

Changed the auto background color to be applied on page load and displayed in the project grid, instead of only applying it when the project tile was clicked.
This way, the graphic project preview is always visible.

The code was only executed once when clicking the project, so it didn't have to be perfectly optimized, but now it required major improvements, since the background color is calculated for multiple graphic projects.
Removed unnecessary parts, simplified the logic and made the calculations into chunks, so there is no lag on page load.

Removed the placeholders and added the actual projects.