v0.12.0 - Contact Form, Nav Link Fixes
Added a contact form.
The data is sent using an external API (so it works on GitHub Pages).
The form has to go through validation checks to be submitted.

There is a check for empty required inputs, invalid email and a honeypot to catch bots.
There is an attempt limit with a timer to further prevent any spam.

If there are any errors / failed validation, a popup with the information will be displayed.
If the API takes too long to respond, a message will be displayed in the popup.
The popup is closed automatically after a set time, by clicking a button or after a short time when scrolling the page.

The submit button is disabled until there's a response from the api or the validation is failed.

Made some minor changes in the NavbarSection component.
The keyframes are not outside the navbar element, so they can be used in other components.
Changed the links to lead to the correct sections / other sites.
Fixed an issue where the nav link scroll to the section would not scroll to the correct place because of the fade-in animation.