v1.0.0 - Design, Accessibility, Bug Fixes
Added dev project links to the GitHub page and the hosted project.

Fixed a bug where the project modals would disappear on page resize.
It was happening because the intersectionObserver from the section fade-in feature would detect that the section is not visible because of the layout shoft on page resize.

Fixed a bug where the project/design modal would not be placed correctly due to the section fade-in animation (changing its position).

Improved the SEO and accessibility.
Converted images from png to webp, resized and compressed them.

Changed the color palette of the website.
Changed the relative color syntax is not widely supported yet, so the colors variants had to be changed to be set manually.

Removed the backdrop-filter:blur from elements.
It was causing poor responsiveness and was incompatible with the opacity fade-in section animation.
Adjusted the gradient background opacity accordingly to compensate for the lack of backdrop blur.

Slightly improved the Firefox support.
Due to Mozilla not supporting major features that have been a standard in the other browsers for years, some things may still not look or work as expected.

Implemented other minor design changes and fixes.