v1.1.0 - Scroll Features, Nav Fixes
Added a scroll-to-top feature on the nav logo click (desktop).

Fixed an issue where the nav link to the "About" section could be accidentally clicked when expanding the navigation on mobile.

Fixed a bug where, if the user clicked the nav link to scroll to the "About" section, the hero section's auto-scroll would be applied, resulting in scrolling all the way to the top of the page.
Additionally, the nav link's target position was adjusted on mobile.

Added a button for scrolling the design projects modal to the top.

Implemented other minor design changes.