v1.2.0 - Custom Cursor
Added a scroll-to-top feature on footer logo click (desktop).
Implemented a simple animation for logos in the nav and footer on hover.

Changed the cursor to a custom one.
It changes its appearance on interactive elements.

Added a blob that follows the cursor with a slight delay, creating a nice atmospheric design language.

Modified certain parts of the navigation scripts to improve performance.