v1.3.0 - Web Design Section & Projects, Performance
Added the Web-Design subsection (Graphic Design) containing app & website designs.
Added an animated web design icon for the subsection tile.

Added more graphic design projects.

Adjusted the project modal code for this subsection so the tiles display the designs correctly.
Implemented a scrolling functionality to the Web-Design project modals, so when the design is long, the user is able to see the whole project.
The design modal displays the design with a custom width (wide for desktop, narrow for mobile designs).

Changed the way the project images are loaded. Instead of the slow load (capping the number of images loaded per second), the images use the lazy loading attribute (some changes had to be made, because the standard HTML attribute alone does not work).
Now, the website performs much better after loading. There is a minor lag / delay when clicking the design tile (for example - Color Palette), but it's unavoidable because of the size of the images that cannot be compressed further.

Fixed a bug where, if the navigation link was clicked while at the top of the page, the auto-scroll feature in the hero section would occasionally not be disabled, resulting in a complete animation.

Implemented lazy loading for the main section components (about, web-dev projects, graphic design, contact).
It greatly improves the loading time, especially on slower devices.

Disabled the section "fade in" animation on mobile devices.
The project now will perform better on less powerful devices that could not keep up with the changing content placement.

Separated the contact form popup into another component to unclutter the code and improve maintainability.

Updated one message in the contact form.
When the API takes too long to respond, the popup message will also include info about the email address, so the user is still able to make contact.

Cleaned up and simplified some code for maintainability.
Implemented other minor changes and improvements (content and accessibility).