v1.3.2 - Bug Fixes
Fixed a bug in the mockup element in the web-dev project section.
When the screenshot of the project was too short, it would not cover the whole mockup frame on mobile.

Fixed a bug where the nav links would not always scroll to the exact position they were supposed to.

Fixed an issue where the footer logo icon would not scroll back to the top on mobile devices.